# lib/api-zod/src/generated/types/incidentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the Schedulync student portal
 * OpenAPI spec version: 0.1.0
 */

export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export const IncidentStatus = {
  submitted: 'submitted',
  in_review: 'in_review',
  resolved: 'resolved',
} as const;